@@ -8,13 +8,10 @@
 </template>
 
 <script lang="ts">
-import { Component, Prop, Vue } from "vue-property-decorator";
-import { State, Getter, Action, Mutation, namespace } from 'vuex-class';
+import { Vue } from "vue-class-component";
+import { Prop } from "vue-property-decorator";
 import { ProductModel, ProductVariantModel, ProductReviewModel } from '../models/ProductModels';
 
-@Component({
-  components: {}
-})
 export default class Cart extends Vue {
 
 
